+++ resources/views/user/index.blade.php
@@ -50,19 +50,25 @@
                                            data-tooltip="Change">
                                             <span class="material-icons">edit</span>
                                         </a>--}}
+                                        @if(hasPermission())
+                                            <a href="{{ route('user.permission', $user->id) }}" class="btn-floating btn-small waves-effect waves-light purple tooltipped" data-position="top"
+                                               data-tooltip="Permission">
+                                                <span class="material-icons">lock_open</span>
+                                            </a>
 
-                                        <a href="javascript:" class="btn-floating btn-small waves-effect waves-light red tooltipped" data-position="top"
-                                           data-tooltip="Delete" onclick="jShowDelete(this)">
-                                            <span class="material-icons">delete</span>
-                                        </a>
+                                            <a href="javascript:" class="btn-floating btn-small waves-effect waves-light red tooltipped" data-position="top"
+                                               data-tooltip="Delete" onclick="jShowDelete(this)">
+                                                <span class="material-icons">delete</span>
+                                            </a>
 
-                                        <div class="delete-form" onclick="jCancelDelete(this)">
-                                            {!! Form::open(['route' => ['user.destroy', $user->id], 'method' => 'DELETE']) !!}
-                                            <h3>You want to delete this. Are you sure?</h3>
-                                            <button type="submit" class="btn red darken-3" onclick="submit_form(this, event)"><span class="material-icons">delete</span> Delete</button>
-                                            <button type="button" class="btn grey" onclick="jCancelDelete(this)"><span class="material-icons">close</span>Cancel</button>
-                                            {!! Form::close() !!}
-                                        </div>
+                                            <div class="delete-form" onclick="jCancelDelete(this)">
+                                                {!! Form::open(['route' => ['user.destroy', $user->id], 'method' => 'DELETE']) !!}
+                                                <h3>You want to delete this. Are you sure?</h3>
+                                                <button type="submit" class="btn red darken-3" onclick="submit_form(this, event)"><span class="material-icons">delete</span> Delete</button>
+                                                <button type="button" class="btn grey" onclick="jCancelDelete(this)"><span class="material-icons">close</span>Cancel</button>
+                                                {!! Form::close() !!}
+                                            </div>
+                                        @endif
                                     </td>
                                 </tr>
                             @endforeach
